requirements.txt
email-validator==1.3.1
fastapi==0.89.1
PyJWT==2.6.0
python-decouple==3.7
uvicorn==0.20.0
motor==3.1.1

